Minor fixes: skill property effectiveRange, summon targeting mistake fixed, 0 items check before selling, being able to set siegeclan type, variable change in l2territory

+++ L2_GameServer_c5/java/net/sf/l2j/gameserver/model/L2Territory.java
@@ -129,8 +129,8 @@ public class L2Territory
 			    p[2] = _z_min+100;p[3] = _z_max;
 			    for( i=0; i<_points.size(); i++){
 				Point p1 = _points.get(i);
-				long dx = p1.x-p[0];
-				long dy = p1.y-p[1];
+				double dx = p1.x-p[0];
+				double dy = p1.y-p[1];
 				double distance = Math.sqrt(dx*dx+dy*dy);
 				if (curdistance == 0 || distance<curdistance){
 				    curdistance = distance;
